Use variant based methods for gpio
stm32f3 provides methods, with which we can access each variant
of a register. No more bit-shifts and -masks are needed.

This also removes a lot off `unsafe` calls, which where needed before
to access a whole register without any bound guards for fields and
flags.

This makes the macro section quite more verbose, because there is
no method with an array index access-point. So we have to be explicit
for every method name we add. `moder0`, `moder1`, ...