serial: Enable DMA only when we actually use it
Instead of always enabling DMA on the USART peripherals, we want to
enable it only when we actually do a transfer. To this end, this commit
renames the `Target` trait mapping DMA targets to their channels to
`OnChannel`. The new `Target` trait now provides methods to start and
stop DMA on the implementing target, which are called when a `Transfer`
starts or stops DMA.