Use correct PACs for stm32f318 and stm32f378
According to the "Supported Devices" table in the stm32f3 crate's README
(https://crates.io/crates/stm32f3) we were using the wrong PAC module
(`stm32f3x8`) for both smt32f318 and stm32f378. Because of that we also
supported some peripherals for these devices that are not actually
available for them.

This commit adds the correct PAC modules for the two devices in question
and fixes up the HAL code to use the correct peripherals.