Fix most GPIO compilation errors
This commit switches to the current nightly version of the stm32f3 PAC
to make use of the fixes regarding the F3 GPIO peripherals.

We also need to adjust the codegen of the GPIO pin mappings a bit to
account for the fact that the F373 GPIO is still special in that the PAC
contains an additional `gpiod` module for it.