Fix Github Actions CI config (#130)
The previous CI executed the `cargo check` command with the `--examples`
flag. The intention was to check both the library itself and the
examples. However, for most MCU versions we don't provide any examples,
they are explicitly feature gated to stm32f303 devices. In these cases
cargo decides that there is nothing to do, so it also doesn't check the
library itself, effectively disabling checking for most MCU families.

We instead want to use the `--all-features` flag. From the cargo docs:
  "Check all targets. This is equivalent to specifying
   --lib --bins --tests --benches --examples."
However, this doesn't work unfortunately since cargo is complaining
about a missing crate "test" for the thumb7em-none-eabihf target. So
instead we explicitly use `--lib --examples`.

This commit also updates the checkout action to v2 since the bug that
existed in this version is officially fixed now.