Add MSRV CI job
As we use intra-doc-links now, which is newly added since 1.48.0
this is a good opportunity for the CI to build specifically for
this version.

This is our MSRV than. However this is only for documentation
purposes and should be increased if newer features are used.
This should not stop anyone from bumping the MSRV.

This complies with
https://github.com/rust-embedded/wg/blob/master/ops/msrv.md