Catch wrong feature selection more robustly
As we are now using build.rs to check the feature selection,
we can now make more sophisticated checks.

Therefor we use build.rs to compare the selected features
and the features in the Cargo.toml (via cargo_metadata),
and verify, that only a single valid device feature is selected.