Notice that some UART peripherals have only PCLK as source
For some devices of the stm32f3 family, the source clock
can not be adjusted and can only be the peripheral clock (PCLK).

As usartXsw() (where X = 1, 2 or 3) would be called directly
in any case, it would not compile for peripheral, which do not have
this flag in their CFGR3 register.

To circumvent this issue, split out the clock logic into it's own macro,
which then can declare, which peripheral has a variable clock source
and which have not.

This conditional check can not be introduced in the former macro
definition, as it is not possible to change the behavior of the macro
depending on a type or name of an identifier.