Support USB on the stm32f303xd/xe devices
The 303xd and 303xe series devices do not have AF14 for pins PA11 and
PA12, so the existing implementation will not work for those devices.

According to an ST document [1], stm32f303xd and xe do not require any
particular mode for the PA11 and PA12 pin, so I've parameterized the
`usb::Peripheral` type using traits so that any mode will be
supported.  I've tried to follow a familiar pattern reminescent of the
gpio alternate function configurations.

[1]: https://www.st.com/resource/en/application_note/dm00260340-migrating-between-stm32f303-and-stm32f302-line-products-stmicroelectronics.pdf