Rework ADC implementation
This also supports all devices now, instead of only f303 variants.

The channel information of the channels where generated automatically,
with help of the stm32cubemx "database" (xml files),
and `xq` a xml wrapper program for `jq`.

As a rework of codegen is needed to support the ADC channel generation
the easier route was taken, including checking in the scripts used
for that.

- Support most important configurations.
- Better OneShot implementation support.
- Interrupt support.